pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tigrinya-tokenizer"
version = "0.1.3"
description = "Robust Tigrinya Byte Pair Encoding tokenizer"
authors = [
  { name="Haben Eyasu", email="habifishe@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.8"

packages = [{ include = "tigrinya_tokenizer" }]
include = [
    "src/tigrinya_tokenizer/configs/*.yaml"
]
dependencies = [
   "tokenizers>=0.13.3",
    "PyYAML>=6.0"
]


license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Cython",
    "Topic :: Text Processing :: Linguistic"
]

[project.urls]
Homepage = "https://github.com/haben-ai/Tigrinya-Tokenizer"
Issues = "https://github.com/haben-ai/Tigrinya-Tokenizer/issues"